move term_of_service_modal to modals folder
