+++ static/js/redux/ui/containers/terms_of_service_modal_container.jsx
@@ -13,7 +13,7 @@ GNU General Public License for more details.
 */
 
 import { connect } from 'react-redux';
-import TermsOfServiceModal from '../terms_of_service_modal';
+import TermsOfServiceModal from '../modals/terms_of_service_modal';
 import { triggerTermsOfServiceModal } from '../../actions/modal_actions';
 import { acceptTOS } from '../../actions/user_actions';
 import { getIsUserInfoIncomplete } from '../../reducers/root_reducer';
